# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BrandResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      logo_url:
        type: string
      name:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
      workflow_state:
        $ref: '#/definitions/dto.StateView'
    type: object
...
  dto.CategoryResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      level:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      slug:
        type: string
    type: object
...
  dto.CompareListResponse:
    properties:
      product_ids:
        items:
          type: integer
        type: array
    type: object
  dto.CompareProductResponse:
    properties:
      allow_backorder:
        type: boolean
      attributes:
        items:
          $ref: '#/definitions/dto.ProductAttributeResponse'
        type: array
      barcode:
        type: string
      brand:
        $ref: '#/definitions/dto.BrandResponse'
      brand_id:
        type: integer
      category:
        $ref: '#/definitions/dto.CategoryResponse'
      category_id:
        type: integer
      channels:
        items:
          type: string
        type: array
      colors:
        items:
          type: string
        type: array
      compare_at_price:
        type: number
      cost:
        type: number
      created_at:
        type: string
      description:
        type: string
      discount_percent:
        type: number
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_digital:
        type: boolean
      is_new:
        type: boolean
      low_stock_threshold:
        type: integer
      meta_description:
        type: string
      meta_title:
        type: string
      name:
        type: string
      price:
        type: number
      published_at:
        type: string
      rating:
        type: number
      return_policy:
        type: string
      reviews_count:
        type: integer
      shipping_info:
        type: string
      sizes:
        items:
          type: string
        type: array
      sku:
        type: string
      slug:
        type: string
      status:
        type: string
      stock:
        type: integer
      store:
        $ref: '#/definitions/dto.ProductStoreSummary'
      store_id:
        type: integer
      store_logo:
        type: string
      store_name:
        type: string
      store_slug:
        type: string
      tags:
        items:
          type: string
        type: array
      track_inventory:
        type: boolean
      updated_at:
        type: string
      visibility:
        description: Publishing extras
        type: string
      warehouse_location:
        type: string
      weight:
        type: number
      workflow_state:
        $ref: '#/definitions/dto.StateView'
    type: object
  dto.CompareRequest:
    properties:
      product_ids:
        items:
          type: integer
        maxItems: 4
        minItems: 2
        type: array
    required:
    - product_ids
    type: object
...
  dto.ProductAttributeResponse:
    properties:
      name:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
...
  dto.ProductStoreSummary:
    properties:
      follower_count:
        type: integer
      id:
        type: integer
      is_verified:
        type: boolean
      location:
        type: string
      logo_url:
        type: string
      name:
        type: string
      rating:
        type: number
      return_policy:
        type: string
      review_count:
        type: integer
      shipping_info:
        type: string
      slug:
        type: string
    type: object
...
  dto.StateView:
    properties:
      code:
        type: string
      color:
        type: string
      id:
        type: integer
      is_final:
        type: boolean
      is_initial:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      text_color:
        type: string
    type: object
...
  dto.SyncCompareRequest:
    properties:
      product_ids:
        items:
          type: integer
        maxItems: 4
        type: array
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@luxe.com
    name: API Support
  description: Production-grade e-commerce backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Shopping Platform API
  version: "1.0"
paths:
...
  /compare:
    get:
      consumes:
      - application/json
      description: Returns the list of product IDs in the compare list (authenticated
        or guest via session_id)
      parameters:
      - description: Session ID for guest users
        in: query
        name: session_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CompareListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth (optional): []
      summary: Get user's compare list
      tags:
      - Compare
    post:
      consumes:
      - application/json
      description: Get product details for comparison (2–4 products)
      parameters:
      - description: Product IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CompareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CompareProductResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Compare products
      tags:
      - Compare
    put:
      consumes:
      - application/json
      description: Save the user's compare list (replaces existing)
      parameters:
      - description: Session ID for guest users
        in: query
        name: session_id
        type: string
      - description: Product IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SyncCompareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth (optional): []
      summary: Sync compare list
      tags:
      - Compare
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"